% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAData_utilities.r
\name{getgdppi}
\alias{getgdppi}
\title{Get a data frame with the GDP price index, either annual or quarterly.}
\usage{
getgdppi(gfreq = "A")
}
\arguments{
\item{gfreq}{"A" (annual) or "Q" (quarterly). Defaults to "A".}
}
\value{
data frame with:

   year (integer), gdppi (double) if freq=="A"

   date (date), gdppi (double) if freq=="Q"
}
\description{
Get a data frame with the GDP price index, either annual or quarterly.
}
\examples{
library("tidyverse")
getgdppi()
getgdppi("Q")
}
