% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAData_utilities.r
\name{getNIPATable}
\alias{getNIPATable}
\title{Get a data frame with the GDP price index, either annual or quarterly.}
\usage{
getNIPATable(gtabnum)
}
\arguments{
\item{gtabnum}{BEA table number to get, as character, with no trailing period. For example, "1.1.1"}
}
\value{
data frame sorted by line number, with:

   vname   BEA variable (series) name
   line    line number
   vdesc   variable description
   tabname full table name
}
\description{
Get a data frame with the GDP price index, either annual or quarterly.
}
\examples{
library("dplyr")
getNIPATable("1.1.1")
getNIPATable("3.3")
}
