% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAData_utilities.r
\name{getNIPATableValue}
\alias{getNIPATableValue}
\title{Return a NIPA table descriptive information as a data frame, with data for 2 latest years.}
\usage{
getNIPATableValue(gtabnum)
}
\arguments{
\item{gtabnum}{BEA table number to get, as character, with no trailing period. For example, "1.1.1"}
}
\value{
data frame sorted by line number, with:

   vname   BEA variable (series) name

   line    line number

   vdesc   variable description

   ynum1   value for 2nd-to-latest year

   ynum2   value for latest year

   tabname full table name
}
\description{
Return a NIPA table descriptive information as a data frame, with data for 2 latest years.
}
\examples{
library("tidyverse")
getNIPATableValue("1.1.1")
getNIPATableValue("3.3")
}
