% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAData_utilities.r
\name{getNIPAvarinfo}
\alias{getNIPAvarinfo}
\title{Get a data frame with information on a variable in a specific BEA table and line number.}
\usage{
getNIPAvarinfo(gtabnum, gline)
}
\arguments{
\item{gtabnum}{BEA table number of desired variable, as character, with no trailing period. For example, "1.1.1".}

\item{gline}{Line number to get, integer.}
}
\value{
A one-row data frame for the variable, with:

   tabnum  BEA table number (just the number, as character)
   line    line number
   vname   BEA variable (series) name
   vdesc   variable description
   tabname full table name
}
\description{
Get a data frame with information on a variable in a specific BEA table and line number.
}
\examples{
library("dplyr")
getNIPAvarinfo("3.3", 1)
getNIPAvarinfo("2.3.2", 4)
}
