% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAData_utilities.r
\name{getNIPAvars}
\alias{getNIPAvars}
\title{Get a data frame with a set of NIPA variables, annual or quarterly.}
\usage{
getNIPAvars(vars, gfreq = "A")
}
\arguments{
\item{vars}{a vector of NIPA variable names}

\item{gfreq}{"A" (annual) or "Q" (quarterly). Defaults to "A".}
}
\value{
data frame with:

   year (integer), vname, value (double), if freq=="A"

   date (date), vname, value (double) if freq=="Q"
}
\description{
Get a data frame with a set of NIPA variables, annual or quarterly.
}
\examples{
library("tidyverse")
vars <- c("A191RX", "A191RG")
getNIPAvars(vars, gfreq="A")
getNIPAvars(vars, gfreq="Q")
}
