% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAData_utilities.r
\name{listtabs}
\alias{listtabs}
\title{Get a list of NIPA.A or NIPA.Q table numbers and descriptions.}
\usage{
listtabs(freq = "A")
}
\arguments{
\item{freq}{"A" (annual) or "Q" (quarterly). Defaults to "A".}
}
\value{
vector with table numbers and descriptions
}
\description{
Get a list of NIPA.A or NIPA.Q table numbers and descriptions.
}
\examples{
library("stringr")
listtabs()
listtabs("Q")
}
