% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipa.a.r
\docType{data}
\name{nipa.a}
\alias{nipa.a}
\title{National Income and Products Accounts, annual.}
\format{Data frame with 1 row per variable per year.
\describe{
\item{tabnum}{NIPA table number, character}
\item{tabname}{NIPA table name, character}
\item{line}{Line number in table, integer}
\item{vdesc}{Variable description, character}
\item{vname}{BEA variable name, character}
\item{year}{Calendar year of data, integer}
\item{value}{Units depend upon the table in question, numeric}
}}
\source{
Bureau of Economic Analysis
}
\usage{
nipa.a
}
\description{
NIPA, annual, various years.
}
\examples{
nipa.a
comment(nipa.a)
}
\keyword{datasets}
