% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipa.au.r
\docType{data}
\name{nipa.au}
\alias{nipa.au}
\title{National Income and Products Accounts, annual, unique variables.}
\format{Data frame with 1 row per variable per year.
\describe{
\item{vname}{BEA variable name, character}
\item{year}{Calendar year of data, integer}
\item{value}{Units depend upon the table in question, numeric}
\item{vdesc}{Variable description, character}
}}
\source{
Bureau of Economic Analysis
}
\usage{
nipa.au
}
\description{
NIPA, annual, unique variables, various years.
}
\examples{
  nipa.au
}
\keyword{datasets}

