% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipa.q.r
\docType{data}
\name{nipa.q}
\alias{nipa.q}
\title{National Income and Products Accounts, quarterly.}
\format{Data frame with 1 row per variable per quarter.
\describe{
\item{tabnum}{NIPA table number, character}
\item{tabname}{NIPA table name, character}
\item{line}{Line number in table, integer}
\item{vdesc}{Variable description, character}
\item{vname}{BEA variable name, character}
\item{date}{First day of quarter, date}
\item{value}{Units depend upon the table in question, numeric}
}}
\source{
Bureau of Economic Analysis
}
\usage{
nipa.q
}
\description{
NIPA, quarterly, various years.
}
\examples{
  nipa.q
}
\keyword{datasets}

