% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipa.qu.r
\docType{data}
\name{nipa.qu}
\alias{nipa.qu}
\title{National Income and Products Accounts, quarterly, unique variables.}
\format{Data frame with 1 row per variable per quarter.
\describe{
\item{vname}{BEA variable name, character}
\item{date}{First day of quarter, date}
\item{value}{Units depend upon the table in question, numeric}
\item{vdesc}{Variable description, character}
}}
\source{
Bureau of Economic Analysis
}
\usage{
nipa.qu
}
\description{
NIPA, quarterly, unique variables, various years.
}
\examples{
require("BEAData")
nipa.qu
comment(nipa.qu)
}
\keyword{datasets}
