% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipa.r
\docType{data}
\name{nipa}
\alias{nipa}
\title{National Income and Products Accounts, annual, quarterly, and monthly, all data.}
\format{
Data frame with 1 row per variable per date.
\describe{
\item{vname}{BEA variable name; character}
\item{date}{first day of year (annual), first day of quarter (quarterly), first day of month (monthly); date}
\item{year}{Calendar year of data; integer}
\item{freq}{frequency: A, Q, or M; character}
\item{value}{Units depend upon the table in question, numeric}
\item{vdesc}{Variable description, character}
}
}
\source{
Bureau of Economic Analysis
}
\usage{
nipa
}
\description{
NIPA, annual, quarterly, and monthly, all data.
}
\examples{
nipa
comment(nipa)
}
\keyword{datasets}
