% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIPAvars.r
\docType{data}
\name{NIPAvars}
\alias{NIPAvars}
\title{Information on variables in the National Income and Products Accounts.}
\format{Data frame with 1 row per variable per BEA table entry
\describe{
\item{vname}{BEA variable name; character}
\item{vdesc}{Variable description, character}
\item{table}{BEA table code description, character}
\item{line}{line number, integer}
\item{tabname}{BEA full table name, character}
\item{tabnum}{BEA table number, character}
}}
\source{
Based on data from Bureau of Economic Analysis
}
\usage{
NIPAvars
}
\description{
Variables in the National Income and Products Accounts.
}
\examples{
NIPAvars
}
\keyword{datasets}
