% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdp.a_all.r
\docType{data}
\name{sgdp.a_all}
\alias{sgdp.a_all}
\title{BEA Regional GDP data all components, annual.}
\format{
Data frame with 1 row per industry per component per year.
\describe{
\item{stabbr}{state abbreviation, factor}
\item{year}{Calendar year, integer}
\item{component}{BEA numeric code, integer}
\item{compname}{BEA name, character}
\item{ind}{BEA industry code, integer}
\item{indclass}{BEA industry classification, character}
\item{indname}{BEA industry name, character}
\item{value}{Amount, numeric}
}
}
\source{
Bureau of Economic Analysis
}
\usage{
sgdp.a_all
}
\description{
Regional, annual all components, various years.
}
\examples{
  sgdp.a_all
}
\keyword{datasets}
