% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdp.q.r
\docType{data}
\name{sgdp.q}
\alias{sgdp.q}
\title{BEA Regional GDP data, quarterly.}
\format{Data frame with 1 row per quarter.
\describe{
\item{stabbr}{state abbreviation, factor}
\item{date}{First day of quarter, date}
\item{gdp}{Nominal gdp, numeric}
\item{rgdp}{Real gdp, numeric}
}}
\source{
Bureau of Economic Analysis
}
\usage{
sgdp.q
}
\description{
Regional, quarterly, various years.
}
\examples{
  sgdp.q
}
\keyword{datasets}
