% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdp_sic.a.r
\docType{data}
\name{sgdp_sic.a}
\alias{sgdp_sic.a}
\title{BEA Regional GDP data, SIC basis, annual.}
\format{
Data frame with 1 row per year.
\describe{
\item{stabbr}{state abbreviation, factor}
\item{year}{Calendar year, integer}
\item{gdp}{Nominal gdp, numeric}
\item{rgdp}{Real gdp, numeric}
}
}
\source{
Bureau of Economic Analysis
}
\usage{
sgdp_sic.a
}
\description{
Regional, annual, SIC basis, various years.
}
\examples{
  sgdp_sic.a
}
\keyword{datasets}
