% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdp_spliced.a.r
\docType{data}
\name{sgdp_spliced.a}
\alias{sgdp_spliced.a}
\title{BEA Regional GDP data, spliced NAICS and SIC basis, annual.}
\format{
Data frame with 1 row per year.
\describe{
\item{stabbr}{state abbreviation, character}
\item{year}{Calendar year, integer}
\item{name}{c("gdp", "rgdp")}
\item{value}{amount, numeric}
}
}
\source{
Bureau of Economic Analysis
}
\usage{
sgdp_spliced.a
}
\description{
Regional, annual, spliced NAICS and SIC basis, various years.
}
\examples{
  sgdp_spliced.a
}
\keyword{datasets}
