% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAData_utilities.r
\name{showtab}
\alias{showtab}
\title{Show values in a NIPA.A or NIPA.Q table, for the latest year or quarter.}
\usage{
showtab(tabnumc, freq = "A")
}
\arguments{
\item{tabnumc}{the table number, as character (e.g., "1.1.1" or "1.1.5")}

\item{freq}{"A" (annual) or "Q" (quarterly). Defaults to "A".}
}
\value{
vector with table numbers and descriptions
}
\description{
Show values in a NIPA.A or NIPA.Q table, for the latest year or quarter.
}
\examples{
library("dplyr")
showtab("1.1.1")
showtab("1.1.1", "Q")
}
