% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spce.a.r
\docType{data}
\name{spce.a}
\alias{spce.a}
\title{BEA Regional state personal consumption expenditure data, annual.}
\format{Data frame with 1 row per year.
\describe{
\item{stabbr}{state abbreviation, factor}
\item{year}{Calendar year, integer}
\item{line}{BEA line number (i.e., expenditure category), integer}
\item{description}{BEA description associated with the line number, character}
\item{value}{State personal consumption expenditures for this line number, nominal, $ billions, numeric}
}}
\source{
Bureau of Economic Analysis
}
\usage{
spce.a
}
\description{
State personal consumption expenditures, annual, various years.
}
\examples{
  spce.a
}
\keyword{datasets}
