% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.a.r
\docType{data}
\name{spi.a}
\alias{spi.a}
\title{BEA Regional state personal income data, annual.}
\format{
Data frame with 1 row per year.
\describe{
\item{stabbr}{state abbreviation, factor}
\item{year}{Calendar year, integer}
\item{pcpi}{Per capita personal income, dollars, numeric}
\item{pop}{State population, thousands, numeric}
\item{spi}{State personal income, nominal, $ billions, numeric}
}
}
\source{
Bureau of Economic Analysis
}
\usage{
spi.a
}
\description{
State personal income and population, annual, various years.
}
\examples{
  spi.a
}
\keyword{datasets}
