% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.a_all.r
\docType{data}
\name{spi.a_all}
\alias{spi.a_all}
\title{BEA Regional Personal Income data all components, annual.}
\format{
Data frame with 1 row per component per year.
\describe{
\item{stabbr}{state abbreviation, factor}
\item{year}{Calendar year, integer}
\item{line}{BEA numeric code, integer}
\item{desc}{BEA name, character}
\item{value}{Amount, numeric}
}
}
\source{
Bureau of Economic Analysis
}
\usage{
spi.a_all
}
\description{
Regional, annual all components, various years.
}
\examples{
  spi.a_all
}
\keyword{datasets}
