% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.q.r
\docType{data}
\name{spi.q}
\alias{spi.q}
\title{BEA Regional state personal income data, quarterly.}
\format{Data frame with 1 row per quarter.
\describe{
\item{stabbr}{state abbreviation, factor}
\item{date}{First day of quarter, date}
\item{line}{BEA line number (i.e., expenditure category), integer}
\item{description}{BEA description associated with the line number, character}
\item{value}{Value, numeric}
}}
\source{
Bureau of Economic Analysis
}
\usage{
spi.q
}
\description{
State personal income components and selected other variables, various years.
}
\examples{
  spi.q
}
\keyword{datasets}
