% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAData-package.r
\docType{package}
\name{BEAData}
\alias{BEAData}
\alias{BEAData-package}
\title{Data from BEA.}
\description{
Currently includes:\cr
\code{\link{nipa.a}} National Income and Products Accounts, annual\cr
\code{\link{nipa.au}} National Income and Products Accounts, annual, unique variables\cr
\code{\link{nipa.q}} National Income and Products Accounts, quarterly\cr
\code{\link{nipa.qu}} National Income and Products Accounts, quarterly, unique variables\cr
\code{\link{sgdp.a}} BEA Regional GDP data, annual\cr
\code{\link{sgdp.a_all}} BEA Regional GDP data all components, annual\cr
\code{\link{sgdp.q}} BEA Regional GDP data, quarterly\cr
\code{\link{sgdp.q_all}} BEA Regional GDP data all components, quarterly\cr
\code{\link{spce.a}} BEA Regional state personal consumption expenditure data, annual\cr
\code{\link{spi.a}} BEA Regional state personal income data, annual\cr
\code{\link{spi.q}} BEA Regional state personal income data, quarterly\cr
}

