% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAData_utilities.r
\name{getNIPATable}
\alias{getNIPATable}
\title{Return a NIPA table descriptive information as a data frame.}
\usage{
getNIPATable(gtabnum)
}
\arguments{
\item{gtabnum}{BEA table number to get, as character, with no trailing period. For example, "1.1.1"}
}
\value{
data frame sorted by line number, with:

   vname   BEA variable (series) name

   line    line number

   vdesc   variable description

   tabname full table name
}
\description{
Return a NIPA table descriptive information as a data frame.
}
\examples{
library("tidyverse")
getNIPATable("1.1.1")
getNIPATable("3.3")
}
